+++ src/CheckTreePicker/test/CheckTreePickerSpec.tsx
@@ -742,4 +742,30 @@ describe('CheckTreePicker', () => {
     rerender(<CheckTreePicker defaultOpen data={data} searchKeyword="Tree" />);
     expect(screen.queryAllByRole('treeitem')).to.have.lengthOf(0);
   });
+
+  describe('Loading state', () => {
+    it('Should display a spinner when loading=true', () => {
+      render(<CheckTreePicker data={data} loading />);
+
+      expect(screen.getByTestId('spinner')).to.exist;
+    });
+
+    it('Should not open menu on click when loading=true', () => {
+      render(<CheckTreePicker data={data} loading />);
+
+      fireEvent.click(screen.getByRole('combobox'));
+
+      expect(screen.queryByRole('listbox')).not.to.exist;
+    });
+
+    it('Should not open menu on Enter key when loading=true', () => {
+      render(<CheckTreePicker data={data} loading />);
+
+      fireEvent.keyDown(screen.getByRole('combobox'), {
+        key: 'Enter'
+      });
+
+      expect(screen.queryByRole('listbox')).not.to.exist;
+    });
+  });
 });