+++ src/Modal/ModalBody.tsx
@@ -5,6 +5,7 @@ import { ModalContext } from './ModalContext';
 import { WithAsProps, RsRefForwardingComponent } from '../@types/common';
 import IconButton from '../IconButton';
 import Close from '@rsuite/icons/Close';
+import DrawerContext from '../Drawer/DrawerContext';
 
 export type ModalBodyProps = WithAsProps;
 
@@ -22,11 +23,13 @@ const ModalBody: RsRefForwardingComponent<'div', ModalBodyProps> = React.forward
     const classes = merge(className, withClassPrefix());
 
     const context = useContext(ModalContext);
-
     const bodyStyles = context?.getBodyStyles?.();
-    return (
-      <Component {...rest} ref={ref} style={{ ...bodyStyles, ...style }} className={classes}>
-        {context?.isDrawer && (
+    const closeButton = useContext(DrawerContext)?.closeButton;
+    let buttonElement: React.ReactNode = null;
+
+    if (closeButton) {
+      buttonElement =
+        typeof closeButton === 'boolean' ? (
           <IconButton
             icon={<Close />}
             appearance="subtle"
@@ -34,7 +37,14 @@ const ModalBody: RsRefForwardingComponent<'div', ModalBodyProps> = React.forward
             className={prefix('close')}
             onClick={context?.onModalClose}
           />
-        )}
+        ) : (
+          closeButton
+        );
+    }
+
+    return (
+      <Component {...rest} ref={ref} style={{ ...bodyStyles, ...style }} className={classes}>
+        {buttonElement}
         {children}
       </Component>
     );